Add function which find all preformatted parts
This function find all preformatted parts and save it to array. Inside
the presormatted text can be some markup which can be handled by
program, so I decided to save it in array and replace it on some symbols
in text to put in the end, when other parts are formatted.