Remove useless part of RegExps
Bold, italic and monospaced regexps have part "[^\*]*", which means any
symbol, except *. I copied bold regexp and modify it to use for italic
and monospaced cases and haven't noticed this because "it works".
Fix this now, so that it does not irritate me. Also it reduces
unnecessary behavior of regexps.