Fix case, where there are no preformatted parts.
When there are no preformatted parts in text, program ends with error,
because array where should be this parts and should be itarated later
are empty and there are no to iterate for. So I added a check, so if
array of pref parts is empty, returns text from findAndSavePref() and
there are no usage of restorePrefs().
It also helps to test converter functions if there ara no pref parts.