Add preformatted error tests and other changes
Write test for catching errors in preformatted text. Cases of test write
as an array of objects, because this is simple and you need just iterate
throught it. But this cases assuming that functuion return something, so
errMDCheck now return errors instead of throwing, and converter now
throw this error.
In the end,I move result of coverting from try/catch in main to separate
invalid markdown file error from markup errors.